Change updateReward mechanism
Now, rewards are applied until last due date and last balance. Only after rewards are applied those values are changed to the next iteration. This ensures that the contracts cannot be exploited to give rewards for an user for moments that it was not eligible for it. In contrast, users can loose rewards if they don't notify deposit right after they stake more veNewo, if they are performing a restake.